+++ src/main/java/com/zea7ot/leetcode/logs.txt
@@ -28,7 +28,7 @@ new
 
 review
 3: 
-4: 
+4: 0249
 5: 
 
 20200813:
