+++ src/main/java/com/zea7ot/leetcode/tags/data_structure/array/by_data_structure/array_string.txt
@@ -17,6 +17,7 @@ References:
 3   1048    https://leetcode.com/problems/longest-string-chain/
 
 4   0030    https://leetcode.com/problems/substring-with-concatenation-of-all-words/
+4   0249    https://leetcode.com/problems/group-shifted-strings/
 4   0474    https://leetcode.com/problems/ones-and-zeroes/
 4   0792    https://leetcode.com/problems/number-of-matching-subsequences/
 4   0916    https://leetcode.com/problems/word-subsets/
