+++ src/main/java/com/zea7ot/leetcode/tags/data_structure/array/n_pointers/summary.txt
@@ -25,6 +25,7 @@ References:
 
 4   0018    https://leetcode.com/problems/4sum/
 4   0030    https://leetcode.com/problems/substring-with-concatenation-of-all-words/
+4   0249    https://leetcode.com/problems/group-shifted-strings/
 4   0287    https://leetcode.com/problems/find-the-duplicate-number/
 4   1477    https://leetcode.com/problems/find-two-non-overlapping-sub-arrays-each-with-target-sum/
 4   1537    https://leetcode.com/problems/get-the-maximum-score/
