+++ src/main/java/com/zea7ot/leetcode/tags/data_structure/map/summary.txt
@@ -29,6 +29,7 @@ References:
 
 4   0030    https://leetcode.com/problems/substring-with-concatenation-of-all-words/
 4   0128    https://leetcode.com/problems/longest-consecutive-sequence/
+4   0249    https://leetcode.com/problems/group-shifted-strings/
 4   0399    https://leetcode.com/problems/evaluate-division/
 4   0442    https://leetcode.com/problems/find-all-duplicates-in-an-array/
 4   0732    https://leetcode.com/problems/my-calendar-iii/
