+++ src/main/java/com/zea7ot/leetcode/tags/data_structure/string/n_pointers/summary.txt
@@ -24,6 +24,7 @@ References:
 
 4   0030    https://leetcode.com/problems/substring-with-concatenation-of-all-words/
 4   0076    https://leetcode.com/problems/minimum-window-substring/
+4   0249    https://leetcode.com/problems/group-shifted-strings/
 4   0727    https://leetcode.com/problems/minimum-window-subsequence/
