+++ src/main/java/com/zea7ot/leetcode/tags/data_structure/string/shifting.txt
@@ -4,10 +4,6 @@ References:
 
 
 ### String
-# Linear Scan / 1 Pointer
-
-1   0067    https://leetcode.com/problems/add-binary/
-
-2   0409    https://leetcode.com/problems/longest-palindrome/
+## Shifting the String
 
 4   0249    https://leetcode.com/problems/group-shifted-strings/
