+++ src/main/java/com/zea7ot/leetcode/tags/data_structure/string/summary.txt
@@ -66,6 +66,7 @@ References:
 4   0097    https://leetcode.com/problems/interleaving-string/
 4   0115    https://leetcode.com/problems/distinct-subsequences/
 4   0139    https://leetcode.com/problems/word-break/
+4   0249    https://leetcode.com/problems/group-shifted-strings/
 4   0267    https://leetcode.com/problems/palindrome-permutation-ii/
 4   0294    https://leetcode.com/problems/flip-game-ii/
 4   0297    https://leetcode.com/problems/serialize-and-deserialize-binary-tree/
